Split budget statement code into gen and custom folders
