+++ src/budget-statement/custom/actions/line-item.ts
@@ -1,10 +1,10 @@
-import { BudgetStatement, LineItemInput } from '../types';
-import { createLineItem } from '../utils';
 import {
     AddLineItemAction,
-    DeleteLineItemAction,
     UpdateLineItemAction,
-} from './types';
+    DeleteLineItemAction,
+} from '../../gen';
+import { LineItemInput, BudgetStatement } from '../types';
+import { createLineItem } from '../utils';
 
 function isEqual(lineItemA: LineItemInput, lineItemB: LineItemInput) {
     return (
@@ -101,6 +101,3 @@ export const deleteLineItemOperation = (
         },
     };
 };
-
-export * from './creators';
-export * from './types';
