+++ src/budget-statement/custom/actions/topup.ts
@@ -1,5 +1,5 @@
+import { RequestTopupAction, TransferTopupAction } from '../../gen';
 import { BudgetStatement } from '../types';
-import { RequestTopupAction, TransferTopupAction } from './types';
 
 export const requestTopupOperation = (
     state: BudgetStatement,
@@ -62,6 +62,3 @@ export const transferTopupOperation = (
         },
     };
 };
-
-export * from './creators';
-export * from './types';
