+++ src/budget-statement/index.ts
@@ -1,6 +1,2 @@
-export * from './account';
-export * from './line-item';
-export * from './status';
-export * from './topup';
-export * from './reducer';
-export * from './utils';
+export * from './gen';
+export * from './custom';
